pyproject.toml
[project]
name = "fastapi-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.6.0",
    "fastapi[standard]>=0.124.4",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
]
